requirements.txt
google-cloud-aiplatform
google-cloud-storage
numpy
tensorflow==2.13.0
tf-agents==0.17.0
tensorboard
tensorboard-plugin-profile
tensorboard-plugin-wit
tensorboard-data-server
tensorflow-io
urllib3
pillow
gsutil
